requirements.txt
fake-bpy-module-latest>=0.0.1